--- services/cgr-engine.go
@@ -1,445 +1,0 @@
-/*
-Real-time Online/Offline Charging System (OCS) for Telecom & ISP environments
-Copyright (C) ITsysCOM GmbH
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>
-*/
-
-package services
-
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"runtime"
-	"runtime/pprof"
-	"sync"
-	"time"
-
-	"github.com/cgrates/birpc"
-	"github.com/cgrates/birpc/context"
-	"github.com/cgrates/cgrates/commonlisteners"
-	"github.com/cgrates/cgrates/config"
-	"github.com/cgrates/cgrates/cores"
-	"github.com/cgrates/cgrates/efs"
-	"github.com/cgrates/cgrates/engine"
-	"github.com/cgrates/cgrates/servmanager"
-	"github.com/cgrates/cgrates/utils"
-	"github.com/cgrates/rpcclient"
-)
-
-func NewCGREngine(stopChan chan struct{}, cfg *config.CGRConfig, sIdxr *servmanager.ServiceIndexer,
-	services []servmanager.Service) *CGREngine {
-	cM := engine.NewConnManager(cfg)
-	caps := engine.NewCaps(cfg.CoreSCfg().Caps, cfg.CoreSCfg().CapsStrategy)
-	shdWg := new(sync.WaitGroup)
-	return &CGREngine{
-		stopChan:   stopChan,
-		cfg:        cfg, // Engine configuration
-		cM:         cM,
-		caps:       caps,  // caps is used to limit RPC CPS
-		shdWg:      shdWg, // wait for shutdown
-		srvManager: servmanager.NewServiceManager(shdWg, cM, cfg, sIdxr, services),
-		srvDep: map[string]*sync.WaitGroup{
-			utils.AccountS:        new(sync.WaitGroup),
-			utils.ActionS:         new(sync.WaitGroup),
-			utils.AdminS:          new(sync.WaitGroup),
-			utils.AnalyzerS:       new(sync.WaitGroup),
-			utils.AsteriskAgent:   new(sync.WaitGroup),
-			utils.AttributeS:      new(sync.WaitGroup),
-			utils.CDRServer:       new(sync.WaitGroup),
-			utils.ChargerS:        new(sync.WaitGroup),
-			utils.CoreS:           new(sync.WaitGroup),
-			utils.DataDB:          new(sync.WaitGroup),
-			utils.DiameterAgent:   new(sync.WaitGroup),
-			utils.DispatcherS:     new(sync.WaitGroup),
-			utils.DNSAgent:        new(sync.WaitGroup),
-			utils.EEs:             new(sync.WaitGroup),
-			utils.EFs:             new(sync.WaitGroup),
-			utils.ERs:             new(sync.WaitGroup),
-			utils.FreeSWITCHAgent: new(sync.WaitGroup),
-			utils.GlobalVarS:      new(sync.WaitGroup),
-			utils.HTTPAgent:       new(sync.WaitGroup),
-			utils.KamailioAgent:   new(sync.WaitGroup),
-			utils.LoaderS:         new(sync.WaitGroup),
-			utils.RadiusAgent:     new(sync.WaitGroup),
-			utils.RateS:           new(sync.WaitGroup),
-			utils.RegistrarC:      new(sync.WaitGroup),
-			utils.ResourceS:       new(sync.WaitGroup),
-			utils.RouteS:          new(sync.WaitGroup),
-			utils.SchedulerS:      new(sync.WaitGroup),
-			utils.SessionS:        new(sync.WaitGroup),
-			utils.SIPAgent:        new(sync.WaitGroup),
-			utils.StatS:           new(sync.WaitGroup),
-			utils.TrendS:          new(sync.WaitGroup),
-			utils.StorDB:          new(sync.WaitGroup),
-			utils.ThresholdS:      new(sync.WaitGroup),
-			utils.TPeS:            new(sync.WaitGroup),
-		},
-		clsCh:      make(chan *commonlisteners.CommonListenerS, 1),
-		anzCh:      make(chan *AnalyzerService, 1),
-		iFilterSCh: make(chan *engine.FilterS, 1),
-	}
-}
-
-type CGREngine struct {
-	stopChan chan struct{}
-
-	cfg *config.CGRConfig
-
-	srvManager *servmanager.ServiceManager
-	srvDep     map[string]*sync.WaitGroup
-	shdWg      *sync.WaitGroup
-	cM         *engine.ConnManager
-
-	caps    *engine.Caps
-	cpuPrfF *os.File
-
-	// services
-	gvS    *GlobalVarS
-	dmS    *DataDBService
-	sdbS   *StorDBService
-	cls    *CommonListenerService
-	anzS   *AnalyzerService
-	coreS  *CoreService
-	cacheS *CacheService
-	ldrs   *LoaderService
-	efs    *ExportFailoverService
-
-	// chans (need to move this as services)
-	clsCh           chan *commonlisteners.CommonListenerS
-	anzCh           chan *AnalyzerService
-	iFilterSCh      chan *engine.FilterS
-	iGuardianSCh    chan birpc.ClientConnector
-	iConfigCh       chan birpc.ClientConnector
-	iServeManagerCh chan birpc.ClientConnector
-	iDispatcherSCh  chan birpc.ClientConnector
-}
-
-func (cgr *CGREngine) GetServDeps() map[string]*sync.WaitGroup {
-	return cgr.srvDep
-}
-
-func (cgr *CGREngine) AddService(service servmanager.Service, connName, apiPrefix string,
-	iConnCh chan birpc.ClientConnector) {
-	cgr.srvManager.AddServices(service)
-	cgr.srvDep[service.ServiceName()] = new(sync.WaitGroup)
-	cgr.cM.AddInternalConn(connName, apiPrefix, iConnCh)
-}
-
-func (cgr *CGREngine) InitServices(setVersions bool) {
-	// init the channel here because we need to pass them to connManager
-	cgr.iServeManagerCh = make(chan birpc.ClientConnector, 1)
-	cgr.iConfigCh = make(chan birpc.ClientConnector, 1)
-	iCoreSv1Ch := make(chan birpc.ClientConnector, 1)
-	iCacheSCh := make(chan birpc.ClientConnector, 1)
-	cgr.iGuardianSCh = make(chan birpc.ClientConnector, 1)
-	iAnalyzerSCh := make(chan birpc.ClientConnector, 1)
-	iCDRServerCh := make(chan birpc.ClientConnector, 1)
-	iAttributeSCh := make(chan birpc.ClientConnector, 1)
-	cgr.iDispatcherSCh = make(chan birpc.ClientConnector, 1)
-	iSessionSCh := make(chan birpc.ClientConnector, 1)
-	iChargerSCh := make(chan birpc.ClientConnector, 1)
-	iThresholdSCh := make(chan birpc.ClientConnector, 1)
-	iStatSCh := make(chan birpc.ClientConnector, 1)
-	iTrendSCh := make(chan birpc.ClientConnector, 1)
-	iRankingSCh := make(chan birpc.ClientConnector, 1)
-	iResourceSCh := make(chan birpc.ClientConnector, 1)
-	iRouteSCh := make(chan birpc.ClientConnector, 1)
-	iAdminSCh := make(chan birpc.ClientConnector, 1)
-	iLoaderSCh := make(chan birpc.ClientConnector, 1)
-	iEEsCh := make(chan birpc.ClientConnector, 1)
-	iRateSCh := make(chan birpc.ClientConnector, 1)
-	iActionSCh := make(chan birpc.ClientConnector, 1)
-	iAccountSCh := make(chan birpc.ClientConnector, 1)
-	iTpeSCh := make(chan birpc.ClientConnector, 1)
-	iEFsCh := make(chan birpc.ClientConnector, 1)
-	iERsCh := make(chan birpc.ClientConnector, 1)
-
-	// initialize the connManager before creating the DMService
-	// because we need to pass the connection to it
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaAnalyzerS), utils.AnalyzerSv1, iAnalyzerSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaAdminS), utils.AdminSv1, iAdminSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaAttributes), utils.AttributeSv1, iAttributeSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaCaches), utils.CacheSv1, iCacheSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaCDRs), utils.CDRsV1, iCDRServerCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaChargers), utils.ChargerSv1, iChargerSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaGuardian), utils.GuardianSv1, cgr.iGuardianSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaLoaders), utils.LoaderSv1, iLoaderSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaResources), utils.ResourceSv1, iResourceSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(rpcclient.BiRPCInternal, utils.MetaSessionS), utils.SessionSv1, iSessionSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaSessionS), utils.SessionSv1, iSessionSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaStats), utils.StatSv1, iStatSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaRankings), utils.RankingSv1, iRankingSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaTrends), utils.TrendSv1, iTrendSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaRoutes), utils.RouteSv1, iRouteSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaThresholds), utils.ThresholdSv1, iThresholdSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaServiceManager), utils.ServiceManagerV1, cgr.iServeManagerCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaConfig), utils.ConfigSv1, cgr.iConfigCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaCore), utils.CoreSv1, iCoreSv1Ch)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaEEs), utils.EeSv1, iEEsCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaRates), utils.RateSv1, iRateSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaDispatchers), utils.DispatcherSv1, cgr.iDispatcherSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaAccounts), utils.AccountSv1, iAccountSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaActions), utils.ActionSv1, iActionSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaTpes), utils.TPeSv1, iTpeSCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaEFs), utils.EfSv1, iEFsCh)
-	cgr.cM.AddInternalConn(utils.ConcatenatedKey(utils.MetaInternal, utils.MetaERs), utils.ErSv1, iERsCh)
-
-	// ServiceIndexer will share service references to all services
-	srvIdxr := servmanager.NewServiceIndexer()
-	cgr.gvS = NewGlobalVarS(cgr.cfg, cgr.srvDep, srvIdxr)
-	cgr.dmS = NewDataDBService(cgr.cfg, cgr.cM, setVersions, cgr.srvDep, srvIdxr)
-	cgr.sdbS = NewStorDBService(cgr.cfg, setVersions, cgr.srvDep, srvIdxr)
-	cgr.cls = NewCommonListenerService(cgr.cfg, cgr.caps, cgr.clsCh, cgr.srvDep, srvIdxr)
-	cgr.anzS = NewAnalyzerService(cgr.cfg, cgr.clsCh,
-		cgr.iFilterSCh, iAnalyzerSCh, cgr.anzCh, cgr.srvDep, srvIdxr)
-
-	cgr.coreS = NewCoreService(cgr.cfg, cgr.caps, cgr.clsCh, iCoreSv1Ch, cgr.anzCh,
-		cgr.cpuPrfF, cgr.shdWg, cgr.srvDep, srvIdxr)
-
-	cgr.cacheS = NewCacheService(cgr.cfg, cgr.dmS, cgr.cM,
-		cgr.clsCh, iCacheSCh, cgr.anzCh, cgr.coreS,
-		cgr.srvDep, srvIdxr)
-
-	dspS := NewDispatcherService(cgr.cfg, cgr.dmS, cgr.cacheS,
-		cgr.iFilterSCh, cgr.clsCh, cgr.iDispatcherSCh, cgr.cM,
-		cgr.anzCh, cgr.srvDep, srvIdxr)
-
-	cgr.ldrs = NewLoaderService(cgr.cfg, cgr.dmS, cgr.iFilterSCh, cgr.clsCh,
-		iLoaderSCh, cgr.cM, cgr.anzCh, cgr.srvDep, srvIdxr)
-
-	cgr.efs = NewExportFailoverService(cgr.cfg, cgr.cM, iEFsCh, cgr.clsCh, cgr.srvDep, srvIdxr)
-
-	cgr.srvManager.AddServices(cgr.gvS, cgr.cls, cgr.coreS, cgr.cacheS,
-		cgr.ldrs, cgr.anzS, dspS, cgr.dmS, cgr.sdbS, cgr.efs,
-		NewAdminSv1Service(cgr.cfg, cgr.dmS, cgr.sdbS, cgr.iFilterSCh, cgr.clsCh,
-			iAdminSCh, cgr.cM, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewSessionService(cgr.cfg, cgr.dmS, cgr.iFilterSCh, cgr.clsCh, iSessionSCh, cgr.cM, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewAttributeService(cgr.cfg, cgr.dmS, cgr.cacheS, cgr.iFilterSCh, cgr.clsCh, iAttributeSCh,
-			cgr.anzCh, dspS, cgr.srvDep, srvIdxr),
-		NewChargerService(cgr.cfg, cgr.dmS, cgr.cacheS, cgr.iFilterSCh, cgr.clsCh,
-			iChargerSCh, cgr.cM, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewRouteService(cgr.cfg, cgr.dmS, cgr.cacheS, cgr.iFilterSCh, cgr.clsCh,
-			iRouteSCh, cgr.cM, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewResourceService(cgr.cfg, cgr.dmS, cgr.cacheS, cgr.iFilterSCh, cgr.clsCh,
-			iResourceSCh, cgr.cM, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewTrendService(cgr.cfg, cgr.dmS, cgr.cacheS, cgr.iFilterSCh, cgr.clsCh,
-			iTrendSCh, cgr.cM, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewRankingService(cgr.cfg, cgr.dmS, cgr.cacheS, cgr.iFilterSCh, cgr.clsCh,
-			iRankingSCh, cgr.cM, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewThresholdService(cgr.cfg, cgr.dmS, cgr.cacheS, cgr.iFilterSCh,
-			cgr.cM, cgr.clsCh, iThresholdSCh, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewStatService(cgr.cfg, cgr.dmS, cgr.cacheS, cgr.iFilterSCh, cgr.clsCh,
-			iStatSCh, cgr.cM, cgr.anzCh, cgr.srvDep, srvIdxr),
-
-		NewEventReaderService(cgr.cfg, cgr.iFilterSCh, cgr.cM, cgr.clsCh, iERsCh, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewDNSAgent(cgr.cfg, cgr.iFilterSCh, cgr.cM, cgr.srvDep, srvIdxr),
-		NewFreeswitchAgent(cgr.cfg, cgr.cM, cgr.srvDep, srvIdxr),
-		NewKamailioAgent(cgr.cfg, cgr.cM, cgr.srvDep, srvIdxr),
-		NewJanusAgent(cgr.cfg, cgr.iFilterSCh, cgr.clsCh, cgr.cM, cgr.srvDep, srvIdxr),
-		NewAsteriskAgent(cgr.cfg, cgr.cM, cgr.srvDep, srvIdxr),                           // partial reload
-		NewRadiusAgent(cgr.cfg, cgr.iFilterSCh, cgr.cM, cgr.srvDep, srvIdxr),             // partial reload
-		NewDiameterAgent(cgr.cfg, cgr.iFilterSCh, cgr.cM, cgr.caps, cgr.srvDep, srvIdxr), // partial reload
-		NewHTTPAgent(cgr.cfg, cgr.iFilterSCh, cgr.clsCh, cgr.cM, cgr.srvDep, srvIdxr),    // no reload
-		NewSIPAgent(cgr.cfg, cgr.iFilterSCh, cgr.cM, cgr.srvDep, srvIdxr),
-
-		NewEventExporterService(cgr.cfg, cgr.iFilterSCh,
-			cgr.cM, cgr.clsCh, iEEsCh, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewCDRServer(cgr.cfg, cgr.dmS, cgr.sdbS, cgr.iFilterSCh, cgr.clsCh, iCDRServerCh,
-			cgr.cM, cgr.anzCh, cgr.srvDep, srvIdxr),
-
-		NewRegistrarCService(cgr.cfg, cgr.cM, cgr.srvDep, srvIdxr),
-
-		NewRateService(cgr.cfg, cgr.cacheS, cgr.iFilterSCh, cgr.dmS,
-			cgr.clsCh, iRateSCh, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewActionService(cgr.cfg, cgr.dmS, cgr.cacheS, cgr.iFilterSCh, cgr.cM,
-			cgr.clsCh, iActionSCh, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewAccountService(cgr.cfg, cgr.dmS, cgr.cacheS, cgr.iFilterSCh,
-			cgr.cM, cgr.clsCh, iAccountSCh, cgr.anzCh, cgr.srvDep, srvIdxr),
-		NewTPeService(cgr.cfg, cgr.cM, cgr.dmS, cgr.clsCh, cgr.srvDep, srvIdxr),
-	)
-
-}
-
-// init will initialize the signal handlers and other functions needed by CGRateS service to run
-func (cgr *CGREngine) init(ctx *context.Context, shtDw context.CancelFunc, flags *CGREngineFlags, vers string) (err error) {
-	cgr.shdWg.Add(1)
-	go cgrSingnalHandler(ctx, shtDw, cgr.cfg, cgr.shdWg)
-
-	var cpuPrfF *os.File
-	if *flags.CpuPrfDir != utils.EmptyString {
-		cpuPath := filepath.Join(*flags.CpuPrfDir, utils.CpuPathCgr)
-		if cpuPrfF, err = cores.StartCPUProfiling(cpuPath); err != nil {
-			return
-		}
-		cgr.cpuPrfF = cpuPrfF
-	}
-
-	if *flags.ScheduledShutdown != utils.EmptyString {
-		var shtDwDur time.Duration
-		if shtDwDur, err = utils.ParseDurationWithNanosecs(*flags.ScheduledShutdown); err != nil {
-			return
-		}
-		cgr.shdWg.Add(1)
-		go func() { // Schedule shutdown
-			tm := time.NewTimer(shtDwDur)
-			select {
-			case <-tm.C:
-				shtDw()
-			case <-ctx.Done():
-				tm.Stop()
-			}
-			cgr.shdWg.Done()
-		}()
-	}
-
-	// init syslog
-	if utils.Logger, err = engine.NewLogger(ctx,
-		utils.FirstNonEmpty(*flags.Logger, cgr.cfg.LoggerCfg().Type),
-		cgr.cfg.GeneralCfg().DefaultTenant,
-		cgr.cfg.GeneralCfg().NodeID,
-		cgr.cM, cgr.cfg); err != nil {
-		return fmt.Errorf("Could not initialize syslog connection, err: <%s>", err)
-	}
-	efs.SetFailedPostCacheTTL(cgr.cfg.EFsCfg().FailedPostsTTL) // init failedPosts to posts loggers/exporters in case of failing
-	utils.Logger.Info(fmt.Sprintf("<CoreS> starting version <%s><%s>", vers, runtime.Version()))
-	cgr.InitServices(*flags.SetVersions)
-	return nil
-}
-
-// startServices will start the services infrastructure
-func (cgr *CGREngine) startServices(ctx *context.Context, shtDw context.CancelFunc, preload string, memProfParams cores.MemoryProfilingParams) (err error) {
-	defer func() {
-		if err != nil {
-			cgr.srvManager.ShutdownServices()
-		}
-	}()
-	cgr.shdWg.Add(1)
-	if err = cgr.gvS.Start(ctx, shtDw); err != nil {
-		cgr.shdWg.Done()
-		return
-	}
-	if cgr.cls.ShouldRun() {
-		cgr.shdWg.Add(1)
-		if err = cgr.cls.Start(ctx, shtDw); err != nil {
-			cgr.shdWg.Done()
-			return
-		}
-	}
-	if cgr.efs.ShouldRun() { // efs checking first because of loggers
-		cgr.shdWg.Add(1)
-		if err = cgr.efs.Start(ctx, shtDw); err != nil {
-			cgr.shdWg.Done()
-			return
-		}
-	}
-	if cgr.dmS.ShouldRun() { // Some services can run without db, ie:  ERs
-		cgr.shdWg.Add(1)
-		if err = cgr.dmS.Start(ctx, shtDw); err != nil {
-			cgr.shdWg.Done()
-			return
-		}
-	}
-	if cgr.sdbS.ShouldRun() {
-		cgr.shdWg.Add(1)
-		if err = cgr.sdbS.Start(ctx, shtDw); err != nil {
-			cgr.shdWg.Done()
-			return
-		}
-	}
-	if cgr.anzS.ShouldRun() {
-		cgr.shdWg.Add(1)
-		if err = cgr.anzS.Start(ctx, shtDw); err != nil {
-			cgr.shdWg.Done()
-			return
-		}
-	} else {
-		cgr.anzCh <- cgr.anzS
-	}
-
-	cgr.shdWg.Add(1)
-	if err = cgr.coreS.Start(ctx, shtDw); err != nil {
-		cgr.shdWg.Done()
-		return
-	}
-	cgr.shdWg.Add(1)
-	if err = cgr.cacheS.Start(ctx, shtDw); err != nil {
-		cgr.shdWg.Done()
-		return
-	}
-	cgr.srvManager.StartServices(ctx, shtDw)
-	// Start FilterS
-	go cgrStartFilterService(ctx, cgr.iFilterSCh, cgr.cacheS.GetCacheSChan(), cgr.cM,
-		cgr.cfg, cgr.dmS)
-
-	cgrInitServiceManagerV1(cgr.iServeManagerCh, cgr.srvManager, cgr.cfg, cgr.clsCh, cgr.anzS)
-	cgrInitGuardianSv1(cgr.iGuardianSCh, cgr.cfg, cgr.clsCh, cgr.anzS)
-	cgrInitConfigSv1(cgr.iConfigCh, cgr.cfg, cgr.clsCh, cgr.anzS)
-
-	if preload != utils.EmptyString {
-		if err = cgrRunPreload(ctx, cgr.cfg, preload, cgr.ldrs); err != nil {
-			return
-		}
-	}
-
-	// Serve rpc connections
-	cgrStartRPC(ctx, shtDw, cgr.cfg, cgr.clsCh, cgr.iDispatcherSCh)
-
-	// TODO: find a better location for this if block
-	if memProfParams.DirPath != "" {
-		if err := cgr.coreS.cS.StartMemoryProfiling(memProfParams); err != nil {
-			utils.Logger.Err(fmt.Sprintf("<%s> %v", utils.CoreS, err))
-		}
-	}
-	return
-}
-
-// Run will run the CGREngine, calling it's init and startServices
-func (cgr *CGREngine) Run(ctx *context.Context, shtDw context.CancelFunc, stopChan chan struct{},
-	flags *CGREngineFlags, vers string, memProfParams cores.MemoryProfilingParams) (err error) {
-	if err = cgr.init(ctx, shtDw, flags, vers); err != nil {
-		return
-	}
-	return cgr.startServices(ctx, shtDw, *flags.Preload, memProfParams)
-}
-
-func (cgr *CGREngine) Stop(pidFile string) {
-	ctx, cancel := context.WithTimeout(context.Background(), cgr.cfg.CoreSCfg().ShutdownTimeout*10)
-	go func() {
-		cgr.shdWg.Wait()
-		cancel()
-	}()
-	<-ctx.Done()
-	if ctx.Err() != context.Canceled {
-		utils.Logger.Err(fmt.Sprintf("<%s> Failed to shutdown all subsystems in the given time",
-			utils.ServiceManager))
-	}
-	if pidFile != utils.EmptyString {
-		if err := os.Remove(pidFile); err != nil {
-			utils.Logger.Warning("Could not remove pid file: " + err.Error())
-		}
-	}
-	if cgr.cpuPrfF != nil && cgr.coreS == nil {
-		pprof.StopCPUProfile()
-		if err := cgr.cpuPrfF.Close(); err != nil {
-			utils.Logger.Err(fmt.Sprintf("<%s> %v", utils.CoreS, err))
-		}
-	}
-
-	// TODO: check if there's any need to manually stop memory profiling.
-	// It should be stopped automatically during CoreS service shutdown.
-
-	utils.Logger.Info("<CoreS> stopped all components. CGRateS shutdown!")
-}
